requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-yasg==1.21.7
-e git+https://github.com/Facturama/facturama-python-sdk.git@7e6a17a61cb01c3868dacfd91734f1d47fbcc025#egg=facturama
h11==0.14.0
idna==3.6
inflection==0.5.1
nose==1.3.7
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.32.0.post1
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.0
